big commit, see commit description
1. Added all the rules:
- adjacent die
- can't be placed against an adjacent die that have same color/value
- discard if one/both condition met.

2. Reworked the `get_dice` function
preventive code, since now, dice can be discarded, I've though that it could mean no more die in dice_set, added code to account for that. However, I don't think this should ever happen. I've backpatched all the code that used the previous implementation.

3. Reworked `place_die` and `placeable`
- They now return different value defined in board.h that help know if die is discarded (used DIE_LOST), place, or not.

4. User can know choose another die
This is to get rid of a softlock bug that happen with the new die-placing conditions. Where user previously could only discard his die.

5.  Update `round_menu`
It now displays the name of the current player that should play a die.

6. Bot update
Bot now check if he can place a die, if he can't check which die the opponent can place and discard it (the most rewarding die for the opponent), if opponent can't play, it discard a random die.

7. After further testing to ensure no bugs, I will remove the `DBG_PRINT` macro